Rename component to SkinTonePicker
